modified scripts and some prefabs
i made a different script to for the boss enemies. Reason, i didnt want the boss to rotate facing the player as the player moves. A script for the boss ship movement is already available.

This is also good because now i can directly affect the boss itself without affecting the enemies with power ups etc because they dont share the same script.